# apps/api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "拾光市集 API",
    "description": "内容种草 + 轻电商 C 端 API。统一响应格式：`{ ok: true, data }` 或 `{ ok: false, error, code }`",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/api", "description": "API 基础路径" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["orders"],
        "summary": "订单列表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "订单列表" } }
      },
      "post": {
        "tags": ["orders"],
        "summary": "创建订单",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fromCart": { "type": "boolean" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": { "type": "string" },
                        "quantity": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "订单创建成功" },
          "400": { "description": "INSUFFICIENT_STOCK" }
        }
      }
    },
    "/orders/{orderId}": {
      "get": {
        "tags": ["orders"],
        "summary": "订单详情",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "订单详情" } }
      }
    },
    "/orders/{orderId}/pay": {
      "post": {
        "tags": ["orders"],
        "summary": "支付订单（Mock）",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "支付成功" },
          "400": { "description": "ORDER_INVALID_STATE" }
        }
      }
    },
    "/orders/{orderId}/cancel": {
      "post": {
        "tags": ["orders"],
        "summary": "取消订单",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "orderId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "取消成功" } }
      }
    },
...
  }
}